Updates for week ending 2021-05-08
Data updates for week ending 2021-05-08.

This update includes two new fields:
* 50% PI
* Abbreviated WIS

Forecast evaluations are based on JHU "truth data", plus back-distributions of error-corrected data.

Data for VA is subject to change for approximately the most recent 4 weeks.

Data for OH is also subject to change as late-reported deaths are back-distributed.